requirements.txt
python-telegram-bot==22.5
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.28.1
